requirements/base.txt
# Django
Django==4.0.8

# Django Rest-Framework
djangorestframework==3.14.0

# Filter
django-filter==21.1

# Postgres
psycopg2-binary==2.9.4

# Environment Variables
django-environ==0.9.0

# Cross-Origin Resource Sharing
django-cors-headers==3.13.0

# Documentation
drf-spectacular==0.24.2

# HTTP library
requests==2.28.1

# Scheduled tasks
celery==5.2.7
django-celery-beat==2.3.0
django-celery-results==2.4.0

# PDF
# weasyprint == 52.5

# Authentication
# djangorestframework-simplejwt==5.2.1